+++ builder/codegen/src/main/java/io/helidon/builder/codegen/BuilderCodegen.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, 2024 Oracle and/or its affiliates.
+ * Copyright (c) 2023, 2025 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -41,6 +41,7 @@ import io.helidon.codegen.spi.CodegenExtension;
 import io.helidon.common.Errors;
 import io.helidon.common.types.AccessModifier;
 import io.helidon.common.types.Annotation;
+import io.helidon.common.types.Annotations;
 import io.helidon.common.types.ElementKind;
 import io.helidon.common.types.TypeInfo;
 import io.helidon.common.types.TypeName;
@@ -159,7 +160,7 @@ class BuilderCodegen implements CodegenExtension {
                     .isStatic(true)
                     .description("Create a new instance from configuration.")
                     .returnType(prototype, "a new instance configured from configuration")
-                    .addParameter(paramBuilder -> paramBuilder.type(Types.COMMON_CONFIG)
+                    .addParameter(paramBuilder -> paramBuilder.type(Types.CONFIG)
                             .name("config")
                             .description("used to configure the new instance"));
             typeArguments.forEach(method::addGenericArgument);
@@ -174,6 +175,26 @@ class BuilderCodegen implements CodegenExtension {
                 }
             }
             classModel.addMethod(method);
+
+            // backward compatibility
+            Method.Builder commonMethod = Method.builder()
+                    .name("create")
+                    .isStatic(true)
+                    .returnType(prototype)
+                    .addParameter(paramBuilder -> paramBuilder.type(Types.COMMON_CONFIG)
+                            .name("config"))
+                    .javadoc(Javadoc.builder()
+                                     .add("Create a new instance from configuration.")
+                                     .returnDescription("a new instance configured from configuration")
+                                     .addParameter("config", "used to configure the new instance")
+                                     .addTag("deprecated", "use {@link #create(" + Types.CONFIG.fqName() + ")}")
+                                     .build())
+                    .addContent("return create(")
+                    .addContent(Types.CONFIG)
+                    .addContentLine(".config(config));")
+                    .addAnnotation(Annotations.DEPRECATED);
+            typeArguments.forEach(commonMethod::addGenericArgument);
+             classModel.addMethod(commonMethod);
         }
     }